+++ src/lib/mrvGL/mrvTimelineWidget.cpp
@@ -815,6 +815,7 @@ namespace mrv
             else
             {
                 p.timelineWidget->setFrameMarkers(markers);
+                p.annotationTimes.clear();
             }
 
             if (_getDrawUpdate(p.timelineWindow) || !p.buffer)